Fix for teacher tests.
