+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/espresso/TestAppManager.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 - present  Instructure, Inc.
+ * Copyright (C) 2022 - present Instructure, Inc.
  *
  *     This program is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -12,26 +12,21 @@
  *
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
+package com.instructure.teacher.ui.espresso
 
-package com.instructure.teacher.utils
-
-import androidx.hilt.work.HiltWorkerFactory
 import androidx.work.Configuration
 import androidx.work.WorkerFactory
-import dagger.hilt.android.HiltAndroidApp
-import javax.inject.Inject
+import com.instructure.teacher.utils.BaseAppManager
 
-@HiltAndroidApp
-class AppManager : BaseAppManager(), Configuration.Provider {
+open class TestAppManager : BaseAppManager() {
 
-    @Inject
-    lateinit var workerFactory: HiltWorkerFactory
+    var workerFactory: WorkerFactory? = null
 
-    override fun getWorkManagerConfiguration(): Configuration =
-        Configuration.Builder()
-            .setWorkerFactory(workerFactory)
-            .build()
-
-    override fun getWorkManagerFactory(): WorkerFactory = workerFactory
+    override fun getWorkManagerFactory(): WorkerFactory {
+        return workerFactory?.let {
+            it
+        } ?: WorkerFactory.getDefaultWorkerFactory()
+    }
 }
