+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/utils/TeacherTest.kt
@@ -18,6 +18,7 @@ package com.instructure.teacher.ui.utils
 
 import android.app.Activity
 import android.view.View
+import androidx.hilt.work.HiltWorkerFactory
 import androidx.test.espresso.UiController
 import androidx.test.espresso.ViewAction
 import androidx.test.espresso.matcher.ViewMatchers
@@ -25,11 +26,15 @@ import com.instructure.canvas.espresso.CanvasTest
 import com.instructure.espresso.InstructureActivityTestRule
 import com.instructure.teacher.BuildConfig
 import com.instructure.teacher.activities.LoginActivity
+import com.instructure.teacher.ui.espresso.TeacherHiltTestApplication
+import com.instructure.teacher.ui.espresso.TeacherHiltTestApplication_Application
 import com.instructure.teacher.ui.pages.*
 import dagger.hilt.android.testing.HiltAndroidRule
 import instructure.rceditor.RCETextEditor
 import org.hamcrest.Matcher
+import org.junit.Before
 import org.junit.Rule
+import javax.inject.Inject
 
 abstract class TeacherTest : CanvasTest() {
 
@@ -38,8 +43,20 @@ abstract class TeacherTest : CanvasTest() {
 
     override val isTesting = BuildConfig.IS_TESTING
 
+    @Inject
+    lateinit var workerFactory: HiltWorkerFactory
+
     @get:Rule(order = 0)
     var hiltRule = HiltAndroidRule(this)
+
+    @Before
+    fun baseSetup() {
+        hiltRule.inject()
+
+        val originalActivity = activityRule.activity
+        val application = originalActivity.application as? TeacherHiltTestApplication_Application
+        application?.workerFactory = workerFactory
+    }
 
     /**
      * Required for auto complete of page objects within tests
