+++ apps/teacher/src/main/java/com/instructure/teacher/utils/BaseAppManager.kt
@@ -20,6 +20,8 @@ import android.content.IntentFilter
 import android.os.Build
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import androidx.work.Configuration
+import androidx.work.WorkerFactory
 import com.google.android.play.core.missingsplits.MissingSplitsManagerFactory
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.heapanalytics.android.Heap
@@ -37,7 +39,7 @@ import com.pspdfkit.PSPDFKit
 import com.pspdfkit.exceptions.InvalidPSPDFKitLicenseException
 import com.pspdfkit.exceptions.PSPDFKitInitializationFailedException
 
-open class BaseAppManager : com.instructure.canvasapi2.AppManager() {
+abstract class BaseAppManager : com.instructure.canvasapi2.AppManager(), Configuration.Provider {
 
     override fun onCreate() {
         if (MissingSplitsManagerFactory.create(this).disableAppIfMissingRequiredSplits()) {
@@ -93,6 +95,13 @@ open class BaseAppManager : com.instructure.canvasapi2.AppManager() {
     override fun performLogoutOnAuthError() {
         TeacherLogoutTask(LogoutTask.Type.LOGOUT).execute()
     }
+
+    override fun getWorkManagerConfiguration(): Configuration =
+        Configuration.Builder()
+            .setWorkerFactory(getWorkManagerFactory())
+            .build()
+
+    abstract fun getWorkManagerFactory(): WorkerFactory
 
     companion object {
         val PREF_FILE_NAME = "teacherSP"
